requirements.txt
# This file was autogenerated by uv via the following command:
#    uv export --format requirements-txt --no-hashes --output-file requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiologger==0.7.0
aiosignal==1.3.2
aiosqlite==0.20.0
annotated-types==0.7.0
anyio==4.8.0
attrs==25.1.0
azure-common==1.1.28
azure-communication-callautomation==1.4.0b1
azure-communication-sms==1.1.0
azure-core==1.32.0
azure-eventgrid==4.21.0
azure-identity==1.19.0
azure-search-documents==11.6.0b4
azure-storage-blob==12.24.1
beautifulsoup4==4.12.3
certifi==2024.12.14
cffi==1.17.1 ; (implementation_name != 'pypy' and os_name == 'nt') or platform_python_implementation != 'PyPy'
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6 ; sys_platform == 'win32'
cryptography==44.0.0
distro==1.9.0
dnspython==2.7.0
email-validator==2.2.0
extruct==0.18.0
fastapi==0.115.8
fastapi-cli==0.0.7
frozenlist==1.5.0
greenlet==3.1.1 ; (python_full_version < '3.14' and platform_machine == 'AMD64') or (python_full_version < '3.14' and platform_machine == 'WIN32') or (python_full_version < '3.14' and platform_machine == 'aarch64') or (python_full_version < '3.14' and platform_machine == 'amd64') or (python_full_version < '3.14' and platform_machine == 'ppc64le') or (python_full_version < '3.14' and platform_machine == 'win32') or (python_full_version < '3.14' and platform_machine == 'x86_64')
h11==0.14.0
html-text==0.6.2
html5lib==1.1
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
isodate==0.7.2
jinja2==3.1.5
jiter==0.8.2
jstyleson==0.0.2
loguru==0.7.3
lxml==5.3.0
lxml-html-clean==0.4.1
markdown-it-py==3.0.0
markupsafe==3.0.2
mdurl==0.1.2
mf2py==2.0.1
msal==1.31.1
msal-extensions==1.2.0
msrest==0.7.1
multidict==6.1.0
oauthlib==3.2.2
openai==1.60.2
outcome==1.3.0.post0
portalocker==2.10.1
propcache==0.2.1
pycparser==2.22 ; (implementation_name != 'pypy' and os_name == 'nt') or platform_python_implementation != 'PyPy'
pydantic==2.10.6
pydantic-core==2.27.2
pygments==2.19.1
pyjwt==2.10.1
pyparsing==3.2.1
pyrdfa3==3.6.4
pysocks==1.7.1
python-dotenv==1.0.1
python-multipart==0.0.20
pywin32==308 ; sys_platform == 'win32'
pyyaml==6.0.2
rapidfuzz==3.12.1
rdflib==7.1.3
recipe-scrapers==15.4.0
redis==5.2.1
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.4
rich-toolkit==0.13.2
rtclient @ https://github.com/Azure-Samples/aoai-realtime-audio-sdk/releases/download/py%2Fv0.5.3/rtclient-0.5.3.tar.gz
selenium==4.28.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
sqlalchemy==2.0.37
starlette==0.45.3
thefuzz==0.22.1
tqdm==4.67.1
trio==0.28.0
trio-websocket==0.11.1
typer==0.15.1
typing-extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0 ; platform_python_implementation != 'PyPy' and sys_platform != 'cygwin' and sys_platform != 'win32'
w3lib==2.3.1
watchfiles==1.0.4
webencodings==0.5.1
websocket-client==1.8.0
websockets==14.2
win32-setctime==1.2.0 ; sys_platform == 'win32'
wsproto==1.2.0
yarl==1.18.3
